requirements.txt
# GENERATED FROM requirements.in.  DO NOT EDIT DIRECTLY.
amqp==5.0.5
asgiref==3.3.1
beautifulsoup4==4.9.3
billiard==3.6.3.0
boto3==1.17.33
botocore==1.20.33
Brotli==1.0.9
bs4==0.0.1
cairocffi==1.2.0
CairoSVG==2.5.2
celery==5.0.5
celery-redbeat==2.0.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.6
cssselect2==0.4.1
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==0.5.0
Django==3.1.7
django-admin-autocomplete-filter==0.7
django-anymail==8.4
django-appconf==1.0.4
django-celery-email==3.0.0
django-colorful==1.3
django-compat==1.0.15
django-cors-headers==3.7.0
django-extensions==3.1.1
django-hijack==2.3.0
django-import-export==2.5.0
django-localflavor==3.0.1
django-object-actions==3.0.1
django-phonenumber-field==5.2.0
django-rest-swagger==2.2.0
django-sslserver==0.22
django-storages==1.11.1
django-widget-tweaks==1.4.8
djangorestframework==3.12.2
djangorestframework-csv==2.1.0
djangorestframework-jwt==1.11.0
ecdsa==0.14.1
et-xmlfile==1.0.1
gunicorn==20.0.4
html2text==2020.1.16
idna==2.10
itypes==1.2.0
Jinja2==2.11.3
jmespath==0.10.0
jsmin==2.2.2
kombu==5.0.2
MarkupPy==1.14
MarkupSafe==1.1.1
mock==4.0.3
oauthlib==3.1.0
odfpy==1.4.1
openapi-codec==1.3.2
openpyxl==3.0.7
phonenumbers==8.12.31
Pillow==8.2.0
prompt-toolkit==3.0.17
psycopg2-binary==2.8.6
pyasn1==0.4.8
pycparser==2.20
PyJWT==1.7.1
python-dateutil==2.8.1
python-decouple==3.4
python-dotenv==0.15.0
python-jose==3.2.0
python-stdnum==1.16
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.4.1
raven==6.10.0
redis==3.5.3
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
s3transfer==0.3.6
simplejson==3.17.2
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.0.2
soupsieve==2.2.1
sqlparse==0.4.1
stripe==2.63.0
tablib==3.0.0
tenacity==7.0.0
tinycss2==1.1.0
unicodecsv==0.14.1
uritemplate==3.0.1
urllib3==1.26.4
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
whitenoise==5.2.0
xlrd==2.0.1
xlwt==1.3.0
yapf==0.31.0
